refactor(item-show-detail): rename component item-detail to item-show-detail
