+++ app/javascript/application.js
@@ -8,7 +8,7 @@ import BaseButton from './components/base-button.vue';
 import BaseCheckbox from './components/base-checkbox.vue';
 import UserSession from './components/user-session.vue';
 import ItemList from './components/item-list.vue';
-import ItemDetail from './components/item-detail.vue';
+import ItemShowDetail from './components/item-show-detail.vue';
 import PurchaseList from './components/purchase-list.vue';
 import './css/application.css';
 
@@ -18,7 +18,7 @@ document.addEventListener('DOMContentLoaded', () => {
       TheHeader,
       UserSession,
       ItemList,
-      ItemDetail,
+      ItemShowDetail,
       PurchaseList,
       BaseNotifications,
     },
