+++ app/javascript/components/item-show-detail.vue
