+++ app/views/app/items/show.html.erb
@@ -6,6 +6,6 @@
     <%= image_tag("back-arrow.svg") %>
     Volver
   </a>
-  <item-detail :item="<%= serialize_resource(@item) %>">
-  </item-detail>
+  <item-show-detail :item="<%= serialize_resource(@item) %>">
+  </item-show-detail>
 </div>
